% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char.num.R
\name{char.num}
\alias{char.num}
\title{A letter counting function}
\usage{
char.num(input_char_str)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
char.num()
}
\keyword{count}
\keyword{string}
