% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charInput.R
\name{charInput}
\alias{charInput}
\title{puts the string array into a pasteable list}
\usage{
charInput(char_array)
}
\arguments{
\item{which}{string array}
}
\description{
generates output for code
}
\examples{
charInput()
}
