% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.num.R
\name{col.num}
\alias{col.num}
\title{A letter counting function}
\usage{
col.num(x)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
col.num()
}
\keyword{count}
\keyword{string}
