% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.df.list.R
\name{combine.df.list}
\alias{combine.df.list}
\title{A letter counting function}
\usage{
combine.df.list(df_list)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
combine.df.list()
}
\keyword{count}
\keyword{string}
