% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossJoin.R
\name{crossJoin}
\alias{crossJoin}
\title{Cross join}
\usage{
crossJoin(...)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
crossJoin()
}
\keyword{count}
\keyword{string}
