% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossJoin.R
\name{crossJoin2}
\alias{crossJoin2}
\title{Cross join}
\usage{
crossJoin2(x, y)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
crossJoin2()
}
\keyword{count}
\keyword{string}
