% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfbind.R
\name{dfbind}
\alias{dfbind}
\title{A letter counting function}
\usage{
dfbind(df_origin, df_new)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
dfbind()
}
\keyword{count}
\keyword{string}
