% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.name.R
\name{file.name}
\alias{file.name}
\title{A letter counting function}
\usage{
file.name(path, original_file_name = "", extension = NA, new_text = "")
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
file.name()
}
\keyword{count}
\keyword{string}
