% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.names.R
\name{get_names}
\alias{get_names}
\title{A letter counting function}
\usage{
get_names(x)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
get_names()
}
\keyword{count}
\keyword{string}
