% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_row.R
\name{insert_row}
\alias{insert_row}
\title{A letter counting function}
\usage{
insert_row(df, insert_below_row, fill = NA)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
insert_row()
}
\keyword{count}
\keyword{string}
