% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justName.R
\name{justName}
\alias{justName}
\title{A letter counting function}
\usage{
justName(file_name)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
justName()
}
\keyword{count}
\keyword{string}
