% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.spatial.R
\name{list.spatial}
\alias{list.spatial}
\title{Lists unique shapefiles}
\usage{
list.spatial(path)
}
\arguments{
\item{what}{path should you list them from}
}
\description{
Lists unique shapefiles
}
\examples{
list.spatial()
}
\keyword{here}
\keyword{is}
\keyword{path}
