% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multmerge.R
\name{multmerge}
\alias{multmerge}
\title{A letter counting function}
\usage{
multmerge(path, header = T, fileEncoding = "")
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
multmerge()
}
\keyword{count}
\keyword{string}
