% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multmerge2.R
\name{multmerge2}
\alias{multmerge2}
\title{A letter counting function}
\usage{
multmerge2(path, read_fn = read.csv)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
multmerge2()
}
\keyword{count}
\keyword{string}
