% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multmerge3.R
\name{multmerge3}
\alias{multmerge3}
\title{A letter counting function}
\usage{
multmerge3(path, read_fn = read.csv, n_files = F, file_path = 1)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
multmerge3()
}
\keyword{count}
\keyword{string}
