% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.rep.R
\name{na.rep}
\alias{na.rep}
\title{A letter counting function}
\usage{
na.rep(data, replacement = 0)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
na.rep()
}
\keyword{count}
\keyword{string}
