% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openwd.R
\name{openwd}
\alias{openwd}
\title{A letter counting function}
\usage{
openwd(a = NULL, justFolder = T)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
openwd()
}
\keyword{count}
\keyword{string}
