% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-from.excel.R
\name{read.excel}
\alias{read.excel}
\title{This reads from clipboard excel cells to dataframe}
\usage{
read.excel(header = TRUE, ...)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
read.excel()
}
\keyword{count}
\keyword{string}
