% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.spatial.R
\name{read.spatial}
\alias{read.spatial}
\title{saves the data frame in a fake file then re-reads it to guess column types}
\usage{
read.spatial(path)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
read.spatial()
}
\keyword{count}
\keyword{string}
