% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{A letter counting function}
\usage{
read_data(path, match_string, data_type = "rds")
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
read_data()
}
\keyword{count}
\keyword{string}
