% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remPunc.R
\name{remPunc}
\alias{remPunc}
\title{A letter counting function}
\usage{
remPunc(str, replacement = "")
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
remPunc()
}
\keyword{count}
\keyword{string}
