% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justPath.R, R/repath.R
\name{repath}
\alias{repath}
\title{A letter counting function}
\usage{
repath()

repath()
}
\arguments{
\item{which}{string would you like to count the letters of}

\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters

This function counts letters
}
\examples{
repath()
repath()
}
\keyword{count}
\keyword{string}
