% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reTable.R
\name{reTable}
\alias{reTable}
\title{A letter counting function}
\usage{
reTable(df_in)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
reTable()
}
\keyword{count}
\keyword{string}
