% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmWhite.R
\name{rmWhite}
\alias{rmWhite}
\title{A letter counting function}
\usage{
rmWhite(x)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
rmWhite()
}
\keyword{count}
\keyword{string}
