% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpt.R
\name{rpt}
\alias{rpt}
\title{repath generation}
\usage{
rpt()
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
rpt()
}
\keyword{count}
\keyword{string}
