% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setpj.R
\name{setpj}
\alias{setpj}
\title{A letter counting function}
\usage{
setpj(project_string = "", newpj = "")
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
setpj()
}
\keyword{count}
\keyword{string}
