% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigFig.R
\name{sigFigInd}
\alias{sigFigInd}
\title{A letter counting function}
\usage{
sigFigInd(num, sf)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
sigFigInd()
}
\keyword{count}
\keyword{string}
