% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_count.R
\name{time_count}
\alias{time_count}
\title{for elements that start and stop dates/times this counts how many are active}
\usage{
time_count(start_times, end_times, time_diff = NA, first = min(start_times),
  last = max(end_times))
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
for elements that start and stop dates/times this counts how many are active
}
\examples{
read.spatial()

}
\keyword{count}
\keyword{string}
