% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titleCase.R
\name{titleCase}
\alias{titleCase}
\title{A letter counting function}
\usage{
titleCase(x)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
titleCase()
}
\keyword{count}
\keyword{string}
