% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.data.R
\name{write.data}
\alias{write.data}
\title{A letter counting function}
\usage{
write.data(data, data_name, folder_name = getwd(), path = "",
  descrition = "", source = list(organisaiton = "", contact = list(name =
  "", email = "")), dateReceived = "YYYY-MM-DD", type = "rds")
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
write.data()


}
\keyword{count}
\keyword{string}
