% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-from.excel.R
\name{write.excel}
\alias{write.excel}
\title{A letter counting function}
\usage{
write.excel(x, row.names = F, col.names = TRUE, ...)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
write.excel()
}
\keyword{count}
\keyword{string}
