% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.spatial.R
\name{write.spatial}
\alias{write.spatial}
\title{A letter counting function}
\usage{
write.spatial(spatial_df, path, overwrite = T)
}
\arguments{
\item{which}{string would you like to count the letters of}
}
\description{
This function counts letters
}
\examples{
write.spatial()

}
\keyword{count}
\keyword{string}
